Fix incompatibility with replay mod, use a character pool
The replay mod uses AudioRecorder objects that are attached to the
skater and board, which have an OnDisabled lifecycle event that closes
a memory stream. Disabling this object when constructing a player from
our player caused the game to crash. I couldn't detect any adverse
effects of letting it enabled (though perhaps there is a small chance
that some audio is sent during this frame.

Before I found the cause of the crash this I was trying a solution using
a pool of controllers so that these can be generated once when the mod
loads and then re-used. However that still caused crashes when the mod
is enabled after the start of the game.

However the object pool might be nice to preserve in some way as constructing
and destroying a MultiplayerPlayerController is relatively expensive.

To make the pool work I had to change it so that only one multiplayerManager
is created so that the pooled controllers can be made before connecting.